Pet Battles: Fix trap animation, captured pet health, and add Newbie achievement
- Use STATUS_CHANGE effect type for trap attempts so the client plays the
  capture crate animation (SET_HEALTH was only updating HP, not triggering
  the visual)
- Captured pets now keep their pre-capture health instead of being set to 0,
  fixing the bug where captured pets appeared dead in the journal and when
  summoned
- Add "Newbie" achievement (ID 7433) SQL hotfix: reach level 3 with a battle
  pet, with proper criteria chain and verified DB2 table hashes
- Fix BattlePetSaveInfo fmt compile error in GrantBattlePetExperience log

Thanks to Celeby for the detailed bug reports!

+++ src/server/game/BattlePets/BattlePetMgr.cpp
@@ -796,7 +796,7 @@ void BattlePetMgr::GrantBattlePetExperience(ObjectGuid guid, uint16 xp, BattlePe
     pet->PacketInfo.Health = pet->PacketInfo.MaxHealth;
 
     TC_LOG_DEBUG("battlepet", "BattlePetMgr::GrantBattlePetExperience: pet {} now level {} xp {} (SaveInfo={})",
-        guid.ToString(), level, pet->PacketInfo.Exp, pet->SaveInfo);
+        guid.ToString(), level, pet->PacketInfo.Exp, static_cast<int32>(pet->SaveInfo));
 
     if (pet->SaveInfo != BATTLE_PET_NEW)
         pet->SaveInfo = BATTLE_PET_CHANGED;
